Renamed selected_countries folder to selected-countries
